obs strategies
