+++ packages/memory/src/processors/observational-memory/__tests__/static-map-cleanup.test.ts
@@ -1,23 +1,22 @@
 import { describe, it, expect, beforeEach } from 'vitest';
 
-import { ObservationalMemory } from '../observational-memory';
+import { BufferingCoordinator } from '../buffering-coordinator';
 
-// Access private static members via `as any` — matches existing test conventions
-const OM = ObservationalMemory as any;
+const BC = BufferingCoordinator as any;
 
 /**
- * Regression tests for OM static map cleanup.
+ * Regression tests for BufferingCoordinator static map cleanup.
  * Ensures that cleanupStaticMaps uses the correct key for reflectionBufferCycleIds.
  */
 
 function clearAllStaticState(): void {
-  OM.asyncBufferingOps.clear();
-  OM.lastBufferedBoundary.clear();
-  OM.lastBufferedAtTime.clear();
-  OM.reflectionBufferCycleIds.clear();
+  BC.asyncBufferingOps.clear();
+  BC.lastBufferedBoundary.clear();
+  BC.lastBufferedAtTime.clear();
+  BC.reflectionBufferCycleIds.clear();
 }
 
-describe('OM static map cleanup', () => {
+describe('BufferingCoordinator static map cleanup', () => {
   beforeEach(() => {
     clearAllStaticState();
   });
@@ -29,31 +28,30 @@ describe('OM static map cleanup', () => {
       const obsBufKey = `obs:${lockKey}`;
       const reflBufKey = `refl:${lockKey}`;
 
-      OM.lastBufferedBoundary.set(obsBufKey, 1000);
-      OM.lastBufferedBoundary.set(reflBufKey, 2000);
-      OM.lastBufferedAtTime.set(obsBufKey, new Date());
-      OM.asyncBufferingOps.set(obsBufKey, Promise.resolve());
-      OM.asyncBufferingOps.set(reflBufKey, Promise.resolve());
-      OM.reflectionBufferCycleIds.set(reflBufKey, 'cycle-abc');
+      BC.lastBufferedBoundary.set(obsBufKey, 1000);
+      BC.lastBufferedBoundary.set(reflBufKey, 2000);
+      BC.lastBufferedAtTime.set(obsBufKey, new Date());
+      BC.asyncBufferingOps.set(obsBufKey, Promise.resolve());
+      BC.asyncBufferingOps.set(reflBufKey, Promise.resolve());
+      BC.reflectionBufferCycleIds.set(reflBufKey, 'cycle-abc');
 
-      const fakeThis = {
-        getLockKey: (_threadId: string, _resourceId?: string | null) => lockKey,
-        getObservationBufferKey: (lk: string) => `obs:${lk}`,
-        getReflectionBufferKey: (lk: string) => `refl:${lk}`,
+      const coordinator = new BufferingCoordinator({
+        observationConfig: { messageTokens: 30000 } as any,
+        reflectionConfig: { observationTokens: 40000 } as any,
         scope: 'thread',
-      };
+      });
 
       // Call cleanupStaticMaps with full cleanup (no activatedMessageIds)
-      ObservationalMemory.prototype['cleanupStaticMaps'].call(fakeThis, 'thread-1', null);
+      coordinator.cleanupStaticMaps('thread-1', null);
 
       // All entries should be removed
-      expect(OM.lastBufferedAtTime.has(obsBufKey)).toBe(false);
-      expect(OM.lastBufferedBoundary.has(obsBufKey)).toBe(false);
-      expect(OM.lastBufferedBoundary.has(reflBufKey)).toBe(false);
-      expect(OM.asyncBufferingOps.has(obsBufKey)).toBe(false);
-      expect(OM.asyncBufferingOps.has(reflBufKey)).toBe(false);
+      expect(BC.lastBufferedAtTime.has(obsBufKey)).toBe(false);
+      expect(BC.lastBufferedBoundary.has(obsBufKey)).toBe(false);
+      expect(BC.lastBufferedBoundary.has(reflBufKey)).toBe(false);
+      expect(BC.asyncBufferingOps.has(obsBufKey)).toBe(false);
+      expect(BC.asyncBufferingOps.has(reflBufKey)).toBe(false);
       // KEY FIX: reflectionBufferCycleIds must be deleted with reflBufKey, not obsBufKey
-      expect(OM.reflectionBufferCycleIds.has(reflBufKey)).toBe(false);
+      expect(BC.reflectionBufferCycleIds.has(reflBufKey)).toBe(false);
     });
 
     it('partial cleanup is a no-op (sealed IDs are now flag-based)', () => {
@@ -61,20 +59,19 @@ describe('OM static map cleanup', () => {
       const obsBufKey = `obs:${lockKey}`;
 
       // Seed some state
-      OM.lastBufferedAtTime.set(obsBufKey, new Date());
+      BC.lastBufferedAtTime.set(obsBufKey, new Date());
 
-      const fakeThis = {
-        getLockKey: () => lockKey,
-        getObservationBufferKey: (lk: string) => `obs:${lk}`,
-        getReflectionBufferKey: (lk: string) => `refl:${lk}`,
+      const coordinator = new BufferingCoordinator({
+        observationConfig: { messageTokens: 30000 } as any,
+        reflectionConfig: { observationTokens: 40000 } as any,
         scope: 'thread',
-      };
+      });
 
       // Partial cleanup: pass activatedMessageIds — should not clear static maps
-      ObservationalMemory.prototype['cleanupStaticMaps'].call(fakeThis, 'thread-1', null, ['msg-1', 'msg-3']);
+      coordinator.cleanupStaticMaps('thread-1', null, ['msg-1', 'msg-3']);
 
       // Static maps should be untouched (partial cleanup only affected sealedMessageIds, which is gone)
-      expect(OM.lastBufferedAtTime.has(obsBufKey)).toBe(true);
+      expect(BC.lastBufferedAtTime.has(obsBufKey)).toBe(true);
     });
   });
 });
